requirements.txt
Flask==2.2.2
Flask_Bcrypt==1.0.1
Flask_Login==0.6.2
Flask_Mail==0.9.1
Flask_SQLAlchemy==2.5.1
Flask_WTF==1.0.1
Pillow==9.2.0
SQLAlchemy==1.4.41
Werkzeug==2.2.2
WTForms==3.0.1
